\name{Flood Flows}
\alias{ardieres}
\title{High Flood Flows of river Ardieres at Beaujeu}
\usage{data(ardieres)}
\description{
  A data frame containing flood discharges over a threshold, in
  units of cubic meters per second, of the Ardieres River at Beaujeu
  (FRANCE), over a period of 33 years and the related date of those
  events.  Unfortunately, the data are   ordered from largest to
  smallest.  The threshold was fixed too low to test functions for the
  threshold selection.}
\format{A vector containing 166 observations.}
\examples{
data(ardieres)
plot(ardieres)
}
\author{Mathieu Ribatet}
\keyword{datasets}
