\name{dens}
\alias{dens}
\title{Density Plot}
\description{
  This is the generic function for density plots.
}
\usage{
dens(fitted, \dots)
}
\arguments{
  \item{fitted}{An object of class ``uvpot'' or ``bvpot''}
  \item{\dots}{Other arguments to be passed for particular methods.}
}
\value{
  A graphic window.
}
\seealso{
  \code{\link{dens.uvpot}}
}
\author{Mathieu Ribatet}
\keyword{hplot}
