\name{Internal functions and methods}
\alias{print.bvpot}
\alias{print.uvpot}
\alias{print.mcpot}
\alias{logLik.pot}
\alias{gpdmoments}
\alias{gpdpwmu}
\alias{gpdpwmb}
\alias{gpdmle}
\alias{gpdpickands}
\alias{gpdmdpd}
\alias{gpdmed}
\alias{qq.gpd}
\alias{pp.gpd}
\alias{dens.gpd}
\alias{retlev.gpd}

\title{Internal functions and methods for the POT package.}
\description{A set of function that should not be used directly by the
  user. For methods, user should usually used the generic functions which
  calls the appropriate method.
}
\author{Mathieu Ribatet}
\keyword{internal}
