\name{retlev}
\alias{retlev}
\title{Return Level Plot}
\description{
  This is the generic function for return level plots.
}
\usage{
retlev(fitted, \dots)
}
\arguments{
  \item{fitted}{An object of class ``uvpot'' or ``bvpot''}
  \item{\dots}{Other arguments to be passed for particular methods.}
}
\value{
  A graphic window.
}
\seealso{
  \code{\link{retlev.uvpot}}, \code{\link{retlev.bvpot}}
}
\author{Mathieu Ribatet}
\keyword{hplot}
