\name{Internal functions and methods}
\alias{gpdmple}
\alias{gpdmgf}
\alias{gpdmoments}
\alias{gpdpwmu}
\alias{gpdpwmb}
\alias{gpdmle}
\alias{gpdpickands}
\alias{gpdmdpd}
\alias{gpdmed}
\alias{gpdlme}
\alias{rbvlog}
\alias{rbvalog}
\alias{rbvnlog}
\alias{rbvanlog}
\alias{rbvmix}
\alias{rbvamix}
\alias{pbvlog}
\alias{pbvalog}
\alias{pbvnlog}
\alias{pbvanlog}
\alias{pbvmix}
\alias{pbvamix}

\title{Internal functions and methods for the POT package.}
\description{A set of functions that should not be used directly by the
  user. For methods, user should usually used the generic functions which
  calls the appropriate method.
}
\author{Mathieu Ribatet}
\keyword{internal}
