\name{Internal functions and methods}
\alias{print.bvpot}
\alias{logLik.pot}
\alias{std.errors.pot}
\title{Internal functions and methods for the POT package.}
\description{A set of function that should not be used directly by the
  user. Fr methods, user should usually used the generic functions which
  calls the appropriate method.
}
\author{Mathieu Ribatet}
\keyword{internal}
