\name{pp}
\alias{pp}
\title{Probability Probability Plot}
\description{
  This is the generic function for probability probability plots.
}
\usage{
pp(fitted, \dots)
}
\arguments{
  \item{fitted}{An object of class ``uvpot'' or ``bvpot''}
  \item{\dots}{Other arguments to be passed for particular methods.}
}
\value{
  A graphic window.
}
\seealso{
  \code{\link{pp.uvpot}}
}
\author{Mathieu Ribatet}
\keyword{hplot}
