\name{qq}
\alias{qq}
\title{Quantile - Quantile Plot}
\description{
  This is the generic function for Quantile - Quantile plots.
}
\usage{
qq(fitted, \dots)
}
\arguments{
  \item{fitted}{An object of class ``uvpot'' or ``bvpot''}
  \item{\dots}{Other arguments to be passed for particular methods.}
}
\value{
  A graphic window.
}
\seealso{
  \code{\link{qq.uvpot}}
}
\author{Mathieu Ribatet}
\keyword{hplot}
